Use preset log error class when message does not match pattern
In the case that a log message can't be broken into "class: message"
format, use the entire log as the message and set an error class based
on the severity of the log.

This change pre-supposes that we make a backend change to how messages
are grouped, taking error title + top line of stacktrace into account
when grouping for unity.