Generate stack trace outside of generator code
It turns out that we cannot generate a stack trace inside of the ienumerator generator code on unity. This will generate a alternative stack trace to use in the case where we are not able to generate one from the exception that we have been past. This is generally when notify is called with an unthrown exception