feat(webgl): Add duration, inForeground and durationInForeground to app metadata
Unity apps deployed to mobile devices as well as webgl would inconsistenly receive data about the
time that the app was running and how long it spent in the foreground, which meant that the averages
were thrown off in the dashboard. This commit implements the recording of the required information
when the notifier is running in a browser.