feat(webgl): Auto detect app version in WebGL
The mobile notifiers auto-detect the app version (e.g. android from the app manifest). The version
number is known in webgl and is added to the Unity metadata tab. This commit adds a LOC that only
runs on webgl to pass the Application.version value through to the "native" JS layer.