Fix session data (#101)
We had an issue where the sessions payload didn't include the same information as the error payload in regards to the application and device data. When the application data differed it caused issues with displaying sessions in the releases tab.

Populate the app and device sections of the payload using the same values sourced from the native notifier as we do when we send an error payload. Additionally the value for version when using the cocoa native notifier did not apply the correct logic to determine the value to appear. This meant that the appVersion set in the Bugsnag configuration wouldn't be used.