fix: Parse Android Java stacktrace frames correctly
Some versions of Unity include a built-in thread uncaught exception
handler, which intercepts Java exceptions and emits them as log messages
if they occur on the main thread. The stacktrace string argument uses a
different format from Unity/C# exceptions, including the actual Java
exception name and message as the first line of the stacktrace.

This change detects android java exceptions by checking the error class
for "AndroidJavaException" then parsing the stacktrace according to the
java format. It fixes a case where the message / actual exception name
is placed into the message.