fix(android): Discard duplicate NDK reports
In the event that a crash occurs from C/C++ code, but Unity Crash
Reporting is enabled, an Exception-level log message is generated and
captured by the bugsnag-unity log handler, effectively re-reporting the
same exception but in an unsymbolicated format as a Java Error.

This changeset adds a check for the presence of the bugsnag crash
reporting library in the log message stacktrace, indicating the report
was previously handled by bugsnag-unity.

A typical symbolicated stack from a signal would start with:

    libunity.{address} NativeRuntimeException::signal_handler()
    libunity.{address} unwind_backtrace_signal_arch()
    bag_handle_signal:{lineno} libbugsnag-ndk.bsg_handle_signal