fix(windows): Allow metadata to be added to the same section
The default behavior of Dictionary is to throw when calling `Add()` with
the same key value. This behavior is handled on platforms with a
NativeClient implementation (Android, iOS, macOS) to instead merge the
values into a single section, though on Windows this behavior was not
overridden. 

In addition, added values which are not dictionaries are coerced into a
dictionary instead of being ignored entirely, improving report
consistency.

Fixes #170.