test: support different .ipa outputs in MR build
Xcode 11 and Xcode 12 output different .ipa names - the product name
"maze_runner" in Xcode 11 and the bundle name "Unity-iPhone" in
Xcode 12. We move it to a versioned path anyway, so a simple fix is
just to wildcard the `mv` command.

Also, removed any existing .ipa's to prevent it picking up previous
builds.