handle null status of client
The client should not be null at this point. I've added an elvis operator to set application state of the client if it isn't null. I think that this might work to some degree, but it doesn't really handle the race condition. Might be better to add a less performative try/catch block instead. or check the client again to see whether it is null and return a log error.